feat: Implement embedding and retrieval controllers with comprehensive API endpoints
- Added embeddingController.py to manage embedding-related API endpoints including status checks, embedding tests, and provider switching.
- Introduced retrievalController.py for document and collection management, including document uploads, searches, and collection operations.
- Created a unified router in router.py to integrate embedding and retrieval functionalities under a single API prefix.
- Developed document_info_generator.py to generate structured metadata for documents using LLMs, with robust error handling and configuration management.
