+++ controller/__init__.py
@@ -7,8 +7,8 @@ from .configController import router as configRouter
 from .workflow.workflowController import router as workflowRouter
 from .node.nodeStateController import router as nodeStateRouter
 from .performanceController import router as performanceRouter
-from .embeddingController import router as embeddingRouter
-from .retrievalController import router as retrievalRouter
+from .rag.embeddingController import router as embeddingRouter
+from .rag.retrievalController import router as retrievalRouter
 from .interactionController import router as interactionRouter
 
 __all__ = [
